# swagger: '2.0'.yaml
swagger: '2.0'
info:
  version: v1
  title: Provider server
  termsOfService: None
basePath: /
paths:
  /Controller/Validate:
    post:
      tags:
        - Controller
      summary: Validate information
      operationId: validate
      consumes:
        - application/json
        - text/json
        - application/json-patch+json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: username
          in: query
          description: Username to login
          required: true
          type: string
        - name: password
          in: query
          description: Password to login
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/User'

  /Controller/GetSuppliers:
    post:
      tags: 
        - Controller
      summary: Gets all suppliers
      operationId: getSuppliers
      produces:
        - text/plain
        - application/json
        - text/json
      responses:
        '200':
          description: ok
          schema:
            type: array
            items:
              $ref: '#/definitions/Page'
              
  /Controller/UpdatePage:
    post:
      tags:
        - Controller
      summary: Update a supplier page
      operationId: updatePage
      consumes:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: page
          in: query
          required: true
          type: string
        - name: description
          in: query
          required: true
          type: string
        - name: location
          in: query
          required: true
          type: string
        - name: contactInformation
          in: query
          required: true
          type: string
      responses:
        '200':
          description: ok
              
  /Controller/GetAllPosts:
    post:
      tags: 
        - Controller
      summary: Gets all posts
      operationId: getAllPosts
      produces:
        - text/plain
        - application/json
        - text/json
      responses:
        '200':
          description: ok
          schema:
            type: array
            items:
              $ref: '#/definitions/Post'
            
  /Controller/AddNoteToSupplier:
    post:
      tags:
        - Controller
      summary: Adds a note to a supplier
      operationId: addNoteToSupplier
      consumes:
        - application/json
        - text/json
        - application/json-patch+json
      parameters:
        - name: supplierName
          in: query
          required: true
          type: string
        - name: editor
          in: query
          required: true
          type: string
        - name: text
          in: query
          required: true
          type: string
      responses:
        '200':
          description: ok
          
  /Controller/CreatePost:
    post:
      tags:
        - Controller
      summary: Creates a post
      operationId: createPost
      consumes:
        - application/json
        - text/json
        - application/json-patch+json
      parameters:
        - name: owner
          in: query
          required: true
          type: string
        - name: title
          in: query
          required: true
          type: string
        - name: description
          in: query
          required: true
          type: string
        - name: type
          in: body
          required: true
          schema:
            $ref: '#/definitions/PostType'
      responses:
        '200':
          description: ok
          schema:
            $ref: '#/definitions/Post'
          
  /Controller/EditPost:
    post:
      tags:
        - Controller
      summary: Edits a posts
      operationId: editPost
      consumes:
        - application/json
        - text/json
        - application/json-patch+json
      parameters:
        - name: post
          in: body
          required: true
          schema:
            $ref: '#/definitions/Post'
        - name: newDescription
          in: query
          required: true
          type: string
        - name: newTitle
          in: query
          required: true
          type: string
      responses:
        '200':
          description: ok
  
  /Controller/DeletePost:
    post:
      tags:
        - Controller
      summary: Deletes a post
      operationId: deletePost
      consumes:
        - application/json
        - text/json
        - application/json-patch+json
      parameters:
        - name: post
          in: body
          required: true
          schema:
            $ref: '#/definitions/Post'
      responses:
        '200':
          description: ok
          
          
  /Controller/CreateProduct:
    post:
      tags:
        - Controller
      summary: Creates a product.
      operationId: createProduct
      consumes:
        - application/json
        - text/json
        - application/json-patch+json
      parameters:
        - name: ProductName
          in: query
          required: true
          type: string
        - name: ChemicalName
          in: query
          required: true
          type: string
        - name: MolWeight
          in: query
          required: true
          type: string
        - name: Description
          in: query
          required: true
          type: string
        - name: Price
          in: query
          required: true
          type: string
        - name: Packaging
          in: query
          required: true
          type: string
        - name: DeliveryTime
          in: query
          required: true
          type: string
        - name: Producer
          in: query
          required: true
          type: string
      responses:
        '200':
          description: ok
          schema:
            $ref: '#/definitions/Product'
         
  /Controller/EditProduct:
    post:
      tags:
        - Controller
      summary: Edits an existing product. 
      operationId: editProduct
      consumes:
        - application/json
        - text/json
        - application/json-patch+json
      parameters:
        - name: product
          in: body
          required: true
          schema:
            $ref: '#/definitions/Product'
        - name: newProductName
          in: query
          required: true
          type: string
        - name: newChemicalName
          in: query
          required: true
          type: string
        - name: newMolWeight
          in: query
          required: true
          type: string
        - name: newDescription
          in: query
          required: true
          type: string
        - name: newPrice
          in: query
          required: true
          type: string
        - name: newPackaging
          in: query
          required: true
          type: string
        - name: newDeliveryTime
          in: query
          required: true
          type: string
      responses:
        '200':
          description: ok


  /Controller/DeleteProduct:
    post:
      tags:
        - Controller
      summary: Deletes an existing product. 
      operationId: deleteProduct
      consumes:
        - application/json
        - text/json
        - application/json-patch+json
      parameters:
        - name: product
          in: body
          required: true
          schema:
            $ref: '#/definitions/Product'
      responses:
        '200':
          description: ok
                    
  /Controller/GetPDF:
    post:
      tags:
        - Controller
      summary: Get the specific PDF for the product 
      operationId: getPDF
      consumes:
        - application/json
        - text/json
        - application/json-patch+json
      produces:
        - text/plain
        - application/json
        - text/json
        - multipart/form-data
      parameters:
        - name: productId
          in: query
          required: true
          type: integer
      responses:
        '200':
          description: ok
          schema:
            type: file

definitions:
  User:
    description: Information about the user
    type: object
    properties:
      username: 
        description: The username of the user
        type: string
      rights:
        description: The rights of the user
        type: string
        enum: [Admin, Provia, Supplier]
  Product:
    description: Informations about a product
    type: object
    properties:
      id:
        description: ID of the product
        type: integer
      producer:
        description: The producer of the product
        type: string
      productName: 
        description: The name of the product
        type: string
      chemicalName:
        description: The chemical name of the product
        type: string
      molWeight:
        description: The mol weight of the product
        type: string
      description:
        description: The description about the product
        type: string
      price:
        description: The price of the product
        type: string
      packaging:
        description: The packaging of the product
        type: string
      deliveryTime:
        description: The delivery time of the product
        type: string
  Page:
    description: Page containing information about a supplier
    type: object
    properties:
      owner:
        description: The owner of the page => The supplier
        type: string
      description:
        description: The description of the page owner
        type: string
      contactInformation:
        description: The contactinformation of the page owner 
        type: string
      location: 
        description: The location of the page owner
        type: string
      products:
        description: The suppliers products
        type: array
        items: 
            $ref: '#/definitions/Product'
      note:
        $ref: '#/definitions/Note'
  Note: 
    description: A note on the supplier page
    type: object
    properties:
      text:
        type: string
      editor:
        type: string
      creationDate:
        type: string
        format: date
  Post: 
    description: A short description of the post.
    type: object
    properties:
      id:
        type: integer
      owner: 
        type: string
      date:
        type: string
      description: 
        type: string
      type:
        $ref: '#/definitions/PostType'
      title: 
        type: string
  PostType:
    type: string
    enum:
      - Warning
      - Request
      - Offer
      - NotAvailable
      
                
        
securityDefinitions: {}